探索レコード最適化: ストレージ効率化とデータモデル統合 (#2)
* CharacterRecord.id を Int32 から UInt8 に変更

キャラクターIDを1〜200の範囲に制限し、型で制約を表現。
採番は最小の未使用IDを割り当て（再利用可能）。

変更点:
- CharacterRecord.id, CharacterEquipmentRecord.characterId
- PartyRecord.memberCharacterIds, PartySnapshot, RuntimeParty
- CharacterSnapshot.id, RuntimeCharacterProgress.id
- BattleActor.partyMemberId, BattleService.Resolution.survivingAllyIds
- ExplorationModels内の全characterId関連フィールド
- 関連サービス・UI（約22ファイル）



* EnemyMaster, ExplorationEventMasterにindex追加

探索レコード最適化Phase 1:
- EnemyMaster.json: 各敵テンプレートにindex (0-9) を追加
- ExplorationEventMaster.json: 各イベントにindex (0-11) を追加
- EnemyDefinition, ExplorationEventDefinitionにindex: UInt16を追加
- SQLiteMasterDataQueries: enemy_index, event_indexカラムの読み取りに対応
- MasterDataGenerator: Schema.swiftとImporters.swiftを更新

* 探索レコード最適化: 4つの@Modelを1つに統合

- ExplorationRunRecord, ExplorationEventRecord, ExplorationEventDropRecord,
  ExplorationBattleLogRecordを1つのExplorationRunRecordに統合
- イベント情報はEventEntry配列をDataとしてエンコードして保存
- UUID識別子からPersistentIdentifier識別子に変更
- String ID → Int Index変換でストレージ効率改善
- イベント発生時刻をEventEntryに追加
- 200件自動パージ機能



* 戦闘ログ数値化: BattleLogEntry → BattleAction/BattleLog

- BattleAction: 数値のみで行動を表現（ストレージ効率化）
- BattleLog: 戦闘ログ全体を保持
- ActionKind: 行動種別を統合したenum
- BattleLogRenderer: 数値ログから表示用BattleLogEntryに変換
- BattleTurnEngine: appendLog → appendAction に全面置換

---------

+++ Epika/Progress/Application/Services/PartyProgressService.swift
@@ -67,7 +67,7 @@ actor PartyProgressService {
         return Self.snapshot(from: party)
     }
 
-    func updatePartyMembers(persistentIdentifier: PersistentIdentifier, memberIds: [Int32]) async throws -> PartySnapshot {
+    func updatePartyMembers(persistentIdentifier: PersistentIdentifier, memberIds: [UInt8]) async throws -> PartySnapshot {
         let context = makeContext()
         let party = try fetchParty(persistentIdentifier: persistentIdentifier, context: context)
         party.memberCharacterIds = memberIds
@@ -106,7 +106,7 @@ actor PartyProgressService {
         return Self.snapshot(from: party)
     }
 
-    func characterIdsInOtherParties(excluding identifier: PersistentIdentifier?) async throws -> Set<Int32> {
+    func characterIdsInOtherParties(excluding identifier: PersistentIdentifier?) async throws -> Set<UInt8> {
         let context = makeContext()
         let descriptor = FetchDescriptor<PartyRecord>()
         let records = try context.fetch(descriptor)
@@ -117,7 +117,7 @@ actor PartyProgressService {
             excludedPartyId = party.id
         }
 
-        var result = Set<Int32>()
+        var result = Set<UInt8>()
         for record in records {
             if record.id != excludedPartyId {
                 result.formUnion(record.memberCharacterIds)
