+++ Epika/Views/States/PartyViewState.swift
@@ -41,7 +41,7 @@ final class PartyViewState {
         try await loadAllParties()
     }
 
-    func updatePartyMembers(party: RuntimeParty, memberIds: [Int32]) async throws {
+    func updatePartyMembers(party: RuntimeParty, memberIds: [UInt8]) async throws {
         _ = try await partyService.updatePartyMembers(persistentIdentifier: party.persistentIdentifier, memberIds: memberIds)
         try await refresh()
     }
